move several functions which replicate ones from postgresql to postgres_ffi crate
